+++ frontend/src/views/sql-editor/EditorCommon/ResultView/SingleResultViewV1.vue
@@ -45,23 +45,13 @@
         >
           {{ $t("sql-editor.visualize-explain") }}
         </NButton>
-        <NDropdown
+        <DataExportButton
           v-if="showExportButton"
-          trigger="hover"
-          :options="exportDropdownOptions"
-          @select="handleExportBtnClick"
-        >
-          <NButton
-            size="small"
-            :loading="isExportingData"
-            :disabled="isExportingData"
-          >
-            <template #icon>
-              <heroicons-outline:download class="h-5 w-5" />
-            </template>
-            {{ t("common.export") }}
-          </NButton>
-        </NDropdown>
+          size="small"
+          :disabled="props.result === null || isEmpty(props.result)"
+          :support-formats="['CSV', 'JSON', 'SQL', 'XLSX']"
+          @export="handleExportBtnClick"
+        />
         <NButton v-else @click="state.showRequestExportPanel = true">
           {{ $t("quick-action.request-export") }}
         </NButton>
@@ -128,8 +118,10 @@ import {
 import { useDebounceFn } from "@vueuse/core";
 import { isEmpty } from "lodash-es";
 import { NInput, NPagination } from "naive-ui";
+import { BinaryLike } from "node:crypto";
 import { computed, reactive, ref } from "vue";
 import { useI18n } from "vue-i18n";
+import { ExportFormat } from "@/components/DataExportButton.vue";
 import RequestExportPanel from "@/components/Issue/panel/RequestExportPanel/index.vue";
 import {
   useInstanceV1Store,
@@ -139,6 +131,7 @@ import {
   useDatabaseV1Store,
   useCurrentUserV1,
 } from "@/store";
+import { useExportData } from "@/store/modules/export";
 import {
   ExecuteConfig,
   ExecuteOption,
@@ -158,7 +151,6 @@ import DataTable from "./DataTable";
 import EmptyView from "./EmptyView.vue";
 import ErrorView from "./ErrorView.vue";
 import { useSQLResultViewContext } from "./context";
-import { useExportData } from "./useExportData";
 
 type LocalState = {
   search: string;
@@ -192,7 +184,7 @@ const instanceStore = useInstanceV1Store();
 const databaseStore = useDatabaseV1Store();
 const currentUserV1 = useCurrentUserV1();
 const dataTable = ref<InstanceType<typeof DataTable>>();
-const { isExportingData, exportData } = useExportData();
+const { exportData } = useExportData();
 const currentTab = computed(() => tabStore.currentTab);
 
 const viewMode = computed((): ViewMode => {
@@ -297,30 +289,10 @@ const pageSize = computed(() => {
   return table.getState().pagination.pageSize;
 });
 
-const exportDropdownOptions = computed(() => [
-  {
-    label: t("sql-editor.download-as-file", { file: "CSV" }),
-    key: "CSV",
-    disabled: props.result === null || isEmpty(props.result),
-  },
-  {
-    label: t("sql-editor.download-as-file", { file: "JSON" }),
-    key: "JSON",
-    disabled: props.result === null || isEmpty(props.result),
-  },
-  {
-    label: t("sql-editor.download-as-file", { file: "SQL" }),
-    key: "SQL",
-    disabled: props.result === null || isEmpty(props.result),
-  },
-  {
-    label: t("sql-editor.download-as-file", { file: "XLSX" }),
-    key: "XLSX",
-    disabled: props.result === null || isEmpty(props.result),
-  },
-]);
-
-const handleExportBtnClick = (format: "CSV" | "JSON" | "SQL" | "XLSX") => {
+const handleExportBtnClick = async (
+  format: ExportFormat,
+  callback: (content: BinaryLike | Blob, format: ExportFormat) => void
+) => {
   const { instanceId, databaseId } = tabStore.currentTab.connection;
   const instance = instanceStore.getInstanceByUID(instanceId).name;
   const database =
@@ -330,7 +302,8 @@ const handleExportBtnClick = (format: "CSV" | "JSON" | "SQL" | "XLSX") => {
   const statement = props.result.statement;
   const admin = tabStore.currentTab.mode === TabMode.Admin;
   const limit = admin ? 0 : RESULT_ROWS_LIMIT;
-  exportData({
+
+  const content = await exportData({
     database,
     instance,
     format,
@@ -338,6 +311,8 @@ const handleExportBtnClick = (format: "CSV" | "JSON" | "SQL" | "XLSX") => {
     limit,
     admin,
   });
+
+  callback(content, format);
 };
 
 const showVisualizeButton = computed((): boolean => {
